requirements.txt
matplotlib==3.7.1
numpy==1.25.2
opencv_contrib_python==4.8.0.76
opencv_python==4.8.0.76
opencv_python_headless==4.9.0.80
PyYAML==6.0.1
tensorflow==2.15.0
